Moved drawing functions into drawing.py.
